Onchain FROST KeyGen Implementation
This PR adds a `FROSTCoordinator` contract for doing onchain FROST
operations. In particular, this PR implements the KeyGen algorithm
from the FROST whitepaper (and assumes that the participants
themselves are responsible for validating the commitments).

We make one notable extension to the algorithm in the whitepaper: in
order to secretly share `f_i(l)` evaluations between the participants,
we use the already submitted "public keys" (i.e. one of the
commitments to `f_i`) to do ECDH and establish a shared secret.

The FROST KeyGen implementation required us to implement some
`secp256k1` curve operations. The "mulmuladd trick" is inherited from
the FROST contracts in the safe-research/safe-frost repository (I can
copy over some of the clarifying comments from there if requested).

Additionally, I added tests for:

- The `secp256k1` implementation
- The FROST KeyGen algorithm in full