Staking Contract `v0.1` (#7)
## TLDR

This is the first version of the Staking contract, which offers the
following features:
- Deposit and Stake any of a particular token (in our case SAFE Token)
for a validator
- The staking contract is ownable, which offers the owner (DAO) to
adjust the withdrawal delay, and to add the validators.
- The first version is a permissioned validator system
- (Design choice, can object) If a validator is removed (the owner has
to wait until a certain delay to remove a validator).
- (Design choice, can object) Except during the initial constructor, the
withdrawal delay could not be higher than the config delay.
- Partial withdrawal of stake is allowed
- Token recovery is built in (with safeToken having different checks)

### TODO

These will be done in a separate PR:
- Testing
- Documentation

### Questions

- Should we include slashing in this version?