Onchain FROST KeyGen Implementation (#8)
This PR adds a `FROSTCoordinator` contract for doing onchain FROST
operations. In particular, this PR implements the KeyGen algorithm from
the FROST whitepaper (and assumes that the participants themselves are
responsible for validating the commitments).

We make one notable extension to the algorithm in the whitepaper: in
order to secretly share `f_i(l)` evaluations between the participants,
we use the already submitted "public keys" (i.e. one of the commitments
to `f_i`) to do ECDH and establish a shared secret.

The FROST KeyGen implementation required us to implement some
`secp256k1` curve operations. The "mulmuladd trick" is inherited from
the FROST contracts in the safe-research/safe-frost repository (I can
copy over some of the clarifying comments from there if requested).

Additionally, I added tests for:

- The `secp256k1` implementation
- The FROST KeyGen algorithm in full

---

An additional note on the coordinator contract implementation: it
currently does not enforce any sort of order on the operations (it is
possible to call `keygenSecretShare` before all participants have called
`keygenCommit`). We did this to reduce the amount of accounting needed
in the coordinator, and require consensus to verify the ceremony was
done correctly before promoting it. We may reconsider this in the future
(as it would allow `KeyGenCompleted` events to be emitted).