Strict Linting and Project Cleanup
We now apply stricter linting rules in CI (this required some changes
to the Staking contract which was throwing some lint errors).

Additionally, I trimmed down the .gitignore file and moved the
contracts build output to the `contracts` subdirectory to match what
we do with the `validators` build output (I originally had everything
under `target/` as I had it setup for a Rust project, and that is
where it puts things).