Trustless KeyGen (#26)
In order to make it so the KeyGen process cannot be DoS'ed by a
malicious participant, we make it fully permissionless. In particular,
the group ID is now deterministically computed based on the group
parameters (with an added `context` for disambiguation). This allows the
epoch rotation KeyGen ceremony to happen without the consensus contract
at all, where the validators themselves are responsible for doing the
KeyGen ceremony and just proposing a rollover to the consensus contract.
In case there is a dishonest participant, the validators can
automatically kick off a new KeyGen ceremony with the dishonest
participant removed from the set.

In order to simplify the validators doing a key generation, a new
`keyGenAndCommit` method was added. It initializes the KeyGen storage
iff it has not already been initialized (instead of reverting like it
does for the `keyGen` function) and providing a DKG commitment. This
helps remove ambiguity from "who is responsible for starting KeyGen"
among the validators, and lets them all kickoff the process without
waiting for one another. The validators themselves are responsible for
determining what the `context` is - tentatively it can be
`address(Consensus) || epoch`.