Verified Signature Share and Challenge (#32)
This PR changes the signature share publishing in the FROST coordinator
to also compute the challenge, instead of relying on the challenge being
part of the signature shares merkle tree.

For us to achieve this, it requires us to:
- Store the message associated with a signature
- Additionally pass in the group commitment `R` when proposing new
participant shares; this also requires us to reflect the group
commitment `R` in the signature shares merkle tree (for this PR, we
reflect it in **every** leaf instead of just once, this simplifies the
onchain logic at the cost of some additional hashing work for the
validators)

As an added bonus, this change allows us to know when a signing ceremony
is complete, and have the contract be able to provide a FROST signature
for a given signature ID. This allows us to simplify the consensus
contract a bit such that it doesn't need to know anything about FROST,
and instead just asks the coordinator "please sign this for me", and at
a later point in time "please verify this signing ceremony for me".

### Glossary

I started calling the subset of participants that are part of a signing
ceremony as a "selection". I chose this word as the subset of
participants is _selected_ from the set of participants by the signature
aggregator, so I thought this was a fitting word. Let me know if you
think I should replace it, and I am open for suggestions.