Use Callback Mechanism to Stage Epochs (#35)
With the new callback mechanism introduced in #34, we can now stage the
epoch as part of the _KeyGen_ completion callback. This allows
validators to only post secret shares and have the epoch rollover
proposal happen automatically with the callback mechanism.

There is a placeholder "TODO" for driving Safe transaction signatures.