[Chore] Biome and TS Configuration Fixes (#43)
Just a small chore PR with some unrelated changes to the storage work on
the build system:

- Fix the Biome configuration to correctly extend the workspace root;
previously it was ignoring the workspace root configuration.
- Removed defaults from Biome configuration (all fields that were
removed were the default, and setting them had no effect).
- Create a separate TSConfig file for outputing the build to `dist/`;
this allows to _only_ output files that are needed to run the validator
and not include tests in the output as well.
- Remove unused TSConfiguration fields (we don't use JSX, nor do we use
decorators, and excluding `node_modules` does nothing, since we don't
include it in the first place).

As a follow up, I will propose bumping up the line length to 120; but I
don't want to do it here and causes a bunch of merge conflicts with the
state machine changes @rmeissner is working on.