Refactor State Machine Specification (#60)
This refactor fixes some of the pending issues that we had previously
identified. Namely:

1. We have a minimum group size (derived from the size of the total
participation set).
2. We skip key generation and signing ceremonies if we do not have a
large enough participant set or signer selection respectively. This adds
a "skipped" state to epoch rollove to indicate that we should wait until
the next epoch to try again. (We do not need an additional skipped state
for signatures, we just remove them from state if they are dropped).
3. Preprocess is now an `Option` instead of a `Map`, this is because the
state machine only allows a single pending nonces chunk per group, and
it simplifies the spec a bit.
4. The `responsible` and `last_participant` are now `Option`s, where
`None` is used to indicate that "everyone is responsible".
5. Old nonces are garbage-collected
6. All cryptography and Ethereum specific code has been moved into their
own modules to not clutter the `Validator` module.
7. Stronger typing on group and signature IDs.

This is a pre-amble to the additions for the new complaint flow. The
refac