Use Logging Framework
This PR adds a logging framework to the project. I chose between
`winston` and `pino`, by far the two most recommended ones on Reddit.

I ended up going with `winston`, even if I had originally opted for
`pino` (as it was smaller, faster, and more widely used according to
NPM.js statistics) because:
- `pino` uses `WebWorker` API to send all logging to a separate thread
  to ensure it doesn't impact the app's performance. This has some
  downsides that affected how we use logging. In particular, not all
  options are serializable which requires a custom transport [1] to
  correctly configure, which doesn't work well with TypeScript [2].
- `pino-pretty`, which we used to make the logging output look nice
  when connected to a TTY, did not work with `vitest`; logs were not
  appearing in `stdout`, and I believe it is something related to how
  `vitest` captures all `stdout` and how `pino-pretty` uses stdout
  redirection [3].

In the end, `winston` just worked, had a much simpler setup, and I
don't think logging performance impact is an issue for us.

[1]: <https://github.com/pinojs/pino-pretty?tab=readme-ov-file#handling-non-serializable-options>
[2]: <https://github.com/pinojs/pino/issues/2203>
[3]: <https://github.com/pinojs/pino-pretty#limitations>