Fix Handling of Own Share in Spec
As @rmeissner pointed out to me, the current spec has different global
states per participant: the last one to emit a `KeyGenSecretShared`
event will eagerly transition into a "confirming" state (as they
already know their share even before it gets included in a block).

In fixing this, I also realized that the spec wasn't correctly
handling the self-share in the first place (as they would try to
verify the secret share from their own event).

In this PR, we fix the spec to correctly:
1. Only register their own share once they see their secret share
   event
2. Do not try to extract, decrypt and verify their a secret share
   from their own event

This should also already be more inline with the changes being made to
the confirmation flow, so we don't need to change the validator as a
result of this change.