Add SQLite Storage Support to Validator Application
Currently, we have implemented various integratinos to use
SQLite-backed storage for persistance on all pieces of validator
state. However, none of this was hooked up to the actual validator
program, and we always used in-memory storage when running it.

This PR adds a new `STORAGE_FILE` option that can be used to use the
persistant SQLite storage for all state (queues, client storage,
consensus state, and machine state).

Additionally, we added a `STORAGE_BACKUP` and `entrypoint.sh` script
that is used in our Docker image. this will automatically on startup
create a backup of the database file before running the validator (if
it exists). This will allow us to replay errors and hopefully make
debugging easier if we encounter any crashes.