Fix Offset for Nonces Commitment
The nonces commitment chunks currently work in a way that allow
participants to commit to nonces partway through a chunk. That is, if
the signing sequence is already at offset `x` (meaning that `x`
signing ceremonies have already started up to signing sequence `x-1`),
then the next `preprocess` nonces commitment will register a chunk
based on that signing sequence. Lets say that the sequence is `1337`,
then it will register chunk `1` with a starting offset of `1337-1024`.
This allows the participant to start revealing nonces and
participating in signing ceremonies right away (that is, starting from
the signing sequence `x`) without having to wait for a new chunk and
recover more quickly. However, because of Wagner's Birthday Attack, we
cannot allow the participant to join signing ceremonies with sequence
`x-1` or earlier, as other signers have have already revealed nonces.

While this logic is useful, there was a small bug when we were
commiting for the _next_ chunk (that is, there was already a nonces
chunk commitment for the current chunk): we didn't set the seqence to
0. This means that if we were on signing sequence `1023` and had
already provided a merkle root for chunk 0 (of which sequence `1023`
belongs), then we would add a new chunk root for chunk 1 **with
starting index `1023`**, making the validators not able to sign any
transactions for the first 1023 offsets of that chunk (i.e. sequences
`1024-2046`). This would make only two sequences from the chunk valid,
even if the participant had registered that chunk's Merkle root before
the chunk started.

This PR fixes that bug, and adds a test to verify the fix.

Note that, while debugging, we had discussed that the `Commitments`
struct `{ uint64 next; mapping(uint64 chunk => Root) chunks }`) is
just an array in disguise, and we could have re-written it as an
array. I decided to keep the current implementation because:
- So we don't need to worry about `uint64 <-> uint256` casting
  semantics and related overflows.
- Avoid innefficients where the `length` would be read multiple
  times (when incrementing while `push()`-ing, to check for
  out-of-bounds access when setting, etc.).